Upgrades wallet impl

+++ script/migrations/61-deploy_protogov_prop.s.sol
@@ -22,9 +22,7 @@ contract KintoMigration61DeployScript is MigrationHelper {
         wallets[0] = _getChainDeployment("KintoWallet-admin");
         uint256[] memory amounts = new uint256[](1);
         amounts[0] = 5e23;
-        selectorAndParams = abi.encodeWithSelector(
-            EngenCredits.setCredits.selector, wallets, amounts
-        );
+        selectorAndParams = abi.encodeWithSelector(EngenCredits.setCredits.selector, wallets, amounts);
         _handleOps(selectorAndParams, address(credits), deployerPrivateKey);
 
         _whitelistApp(_getChainDeployment("EngenGovernance"), deployerPrivateKey, true);
@@ -33,7 +31,7 @@ contract KintoMigration61DeployScript is MigrationHelper {
         bytes[] memory data;
         uint256[] memory values;
         string memory description;
-        uint proposalId;
+        uint256 proposalId;
 
         targets = new address[](1);
         targets[0] = address(_getChainDeployment("Counter"));
@@ -43,34 +41,26 @@ contract KintoMigration61DeployScript is MigrationHelper {
         values[0] = 0;
         description = "ENIP:1 - Kinto Constitution";
 
-
-        selectorAndParams = abi.encodeWithSelector(
-            Governor.propose.selector, targets, values, data, description
-        );
+        selectorAndParams = abi.encodeWithSelector(Governor.propose.selector, targets, values, data, description);
         _handleOps(selectorAndParams, address(governance), deployerPrivateKey);
 
-        proposalId =  governance.hashProposal(targets, values, data, keccak256(bytes(description)));
+        proposalId = governance.hashProposal(targets, values, data, keccak256(bytes(description)));
         require(governance.state(proposalId) == IGovernor.ProposalState.Pending);
-        console.log('Proposal ID 1:', proposalId);
+        console.log("Proposal ID 1:", proposalId);
 
         description = "ENIP:2 - The Kinto Token";
-        selectorAndParams =
-            abi.encodeWithSelector(Governor.propose.selector, targets, values, data, description);
+        selectorAndParams = abi.encodeWithSelector(Governor.propose.selector, targets, values, data, description);
         _handleOps(selectorAndParams, address(governance), deployerPrivateKey);
-        proposalId =  governance.hashProposal(targets, values, data, keccak256(bytes(description)));
+        proposalId = governance.hashProposal(targets, values, data, keccak256(bytes(description)));
         require(governance.state(proposalId) == IGovernor.ProposalState.Pending);
-        console.log('Proposal ID 2:', proposalId);
-
+        console.log("Proposal ID 2:", proposalId);
 
         description = "ENIP:3 - The Mining Program";
-        selectorAndParams = abi.encodeWithSelector(
-            Governor.propose.selector, targets, values, data, description
-        );
+        selectorAndParams = abi.encodeWithSelector(Governor.propose.selector, targets, values, data, description);
         _handleOps(selectorAndParams, address(governance), deployerPrivateKey);
 
-        proposalId =  governance.hashProposal(targets, values, data, keccak256(bytes(description)));
+        proposalId = governance.hashProposal(targets, values, data, keccak256(bytes(description)));
         require(governance.state(proposalId) == IGovernor.ProposalState.Pending);
-        console.log('Proposal ID 3:', proposalId);
-
+        console.log("Proposal ID 3:", proposalId);
     }
 }
